--- a/pymatbridge/matlab/util/json_v0.2.2/json/json_dump.m
+++ b/pymatbridge/matlab/util/json_v0.2.2/json/json_dump.m
@@ -101,10 +101,11 @@ function obj = dump_data_(value, options)
   elseif ~isscalar(value)
     obj = javaObject('org.json.JSONArray');
 
-    if strcmp(class(value), 'double') && ~iscell(value)
-        % encode double arrays as a struct
+    if isnumeric(value) && isreal(value)
+        % encode arrays as a struct
         double_struct = struct;
         double_struct.ndarray = 1;
+        value = double(value);
         double_struct.data = base64encode(typecast(value(:), 'uint8'));
         double_struct.shape = base64encode(typecast(size(value), 'uint8'));
         obj = dump_data_(double_struct, options);