… pymatbridge/tests/test_get_variable.py
… pymatbridge/tests/test_magic.py