@blink1073 2015-01-17
I'd prefer this to the create-a-file approach, since it is all in memory, and does not require you to be on the same machine to share the data.