@arokem 2015-01-17
Sounds great. I will not have time to look at it before Monday, but I will
do it then.

On Saturday, January 17, 2015, Steven Silvester <notifications@github.com>
wrote:

> All set!
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/arokem/python-matlab-bridge/pull/103#issuecomment-70373823>
> .
>