@blink1073 2015-01-17
This was also an exercise in grokking the code base.  FYI, my Matlab license expires in ~40 days.